Use the CUDA sticky errors plugin for scikit-learn examples as well
